HTBHF-2507 move session accessors to top level of flow control directory
