+++ src/web/routes/application/flow-control/session-accessors/index.js
@@ -1,0 +1,13 @@
+const {
+  setNextAllowedPathInSession,
+  setStateInSession,
+  getNextAllowedPathFromSession,
+  getStateFromSession
+} = require('./session-accessors')
+
+module.exports = {
+  setNextAllowedPathInSession,
+  setStateInSession,
+  getNextAllowedPathFromSession,
+  getStateFromSession
+}
